Move usePostMessageListener to shared/
